backend/bot/requirements.txt
python-telegram-bot
openai
manim
python-dotenv
google-api-python-client
google-auth-oauthlib
google-auth-httplib2